refactor: Refactored crypto.py. Tests are known not to work because of issue import telegram_bot.py
